--- Server/src/main/java/ForMZ/Server/domain/like/postLike/controller/PostLikeController.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.like.postLike.controller;
-
-public class PostLikeController {
-}
