--- Server/src/main/java/ForMZ/Server/domain/like/postLike/dto/PostLikeReq.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.like.postLike.dto;
-
-public class PostLikeReq {
-}
