--- Server/src/main/java/ForMZ/Server/domain/like/postLike/exception/PostLikeNotFoundException.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.like.postLike.exception;
-
-public class PostLikeNotFoundException {
-}
