--- Server/src/main/java/ForMZ/Server/domain/like/postLike/mapper/PostLikeMapper.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.like.postLike.mapper;
-
-public interface PostLikeMapper {
-}
