--- Server/src/main/java/ForMZ/Server/domain/like/postLike/repository/PostLikeRepository.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.like.postLike.repository;
-
-public interface PostLikeRepository {
-}
