--- Server/src/main/java/ForMZ/Server/domain/like/postLike/service/PostLikeService.java
@@ -1,4 +1,0 @@
-package ForMZ.Server.domain.like.postLike.service;
-
-public class PostLikeService {
-}
