+++ Server/src/main/java/ForMZ/Server/domain/postLike/controller/PostLikeController.java
@@ -1,0 +1,4 @@
+package ForMZ.Server.domain.postLike.controller;
+
+public class PostLikeController {
+}
