+++ Server/src/main/java/ForMZ/Server/domain/postLike/dto/PostLikeReq.java
@@ -1,0 +1,4 @@
+package ForMZ.Server.domain.postLike.dto;
+
+public class PostLikeReq {
+}
