+++ Server/src/main/java/ForMZ/Server/domain/postLike/exception/PostLikeNotFoundException.java
@@ -1,0 +1,4 @@
+package ForMZ.Server.domain.postLike.exception;
+
+public class PostLikeNotFoundException {
+}
