+++ Server/src/main/java/ForMZ/Server/domain/postLike/mapper/PostLikeMapper.java
@@ -1,0 +1,4 @@
+package ForMZ.Server.domain.postLike.mapper;
+
+public interface PostLikeMapper {
+}
