+++ Server/src/main/java/ForMZ/Server/domain/postLike/repository/PostLikeRepository.java
@@ -1,0 +1,4 @@
+package ForMZ.Server.domain.postLike.repository;
+
+public interface PostLikeRepository {
+}
