+++ Server/src/main/java/ForMZ/Server/domain/postLike/service/PostLikeService.java
@@ -1,0 +1,4 @@
+package ForMZ.Server.domain.postLike.service;
+
+public class PostLikeService {
+}
